Refactor main.cpp and move to Gameboy class
